+++ qcop/adapters/utils.py
@@ -245,7 +245,7 @@ def construct_provenance(
     return Provenance(
         program=program,
         program_version=version,
-        scratch_dir=str(scratch_dir),
+        scratch_dir=scratch_dir,
         wall_time=round(wall_time, 6),
         # TODO: Possibly add psutil.virtual_memory().total and then wrap these calls in
         # @lru_cache so that for quick calculations we don't have to call these